+++ packages/prisma/prisma/migrations/20260430173523_add_smtp_provider_field/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE "SmtpAccount" ADD COLUMN     "provider" TEXT NOT NULL DEFAULT 'smtp',
+ADD COLUMN     "sesConfigSet" TEXT;